requirements.txt
mesa==3.2.0
pyyaml>=6.0
websockets>=12.0
python-dotenv>=1.0
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12